Title: "백엔드 리펙토링 2차 수정 완료"

+++ src/main/java/com/douzone/surveymanagement/statistics/controller/SelectionController.java
@@ -44,7 +44,7 @@ public class SelectionController {
 //    }
 
     @GetMapping("/resultall")
-    public ResponseEntity<CommonResponse>  findSelectionListAll(@RequestParam(value = "surveyno") int surveyNo){
+    public ResponseEntity<CommonResponse>  findSelectionListAll(@RequestParam(value = "surveyno") long surveyNo){
 
         Authentication authenticationCheck = SecurityContextHolder.getContext().getAuthentication();
         HttpHeaders headers = new HttpHeaders();
